refactor: define ItemSearchOption related class inside ItemSearchOption
